+++ android/feature/nowplaying/src/test/snapshots/images/ly.david.musicsearch.android.feature.nowplaying_NowPlayingHistoryCardTest_default[NIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:899461bdc7649292af374549d65b63a0f3014eaa2719a179a6b4318c7b0614bb
-size 16286
+oid sha256:55ef9bf9a6746ce0a0c8181164be5ba9ed7c145243fb194e27387d1ee6ad53c9
+size 16131
