+++ android/feature/spotify/src/test/snapshots/images/ly.david.musicsearch.android.feature.spotify.history_SpotifyHistoryCardTest_default[NOTNIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:0b564bfebe545076e0364db089bb8b4dec273e2ca43fd01ce168b85ee3f943e1
-size 17212
+oid sha256:daae7d8b2440bcb01616c2e18e10ee2a2a0cc9e6a11345caa4461e8b6fc05e78
+size 17317
