+++ android/feature/spotify/src/test/snapshots/images/ly.david.musicsearch.android.feature.spotify.history_SpotifyHistoryUiTest_default[NIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:1ba34897fd4ed8aca8fb875972eafe71f1620c21165066fd5506d2a9bf089bf8
-size 38987
+oid sha256:cf5e735e6ff0341bb3bfdf936a8c11c18ac3b7569d6fb4d7d91f446be1461d77
+size 38949
