+++ android/feature/spotify/src/test/snapshots/images/ly.david.musicsearch.android.feature.spotify.history_SpotifyHistoryUiTest_default[NOTNIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:ee57aff318291ae46eb571415c7af48f58e8ff48ed809fd39a7b6caea1316ac3
-size 39528
+oid sha256:507de59eb6a523abd6d43a3544334f01d85ac039687ac930929eca99566c22e1
+size 39554
