+++ shared/feature/history/src/androidUnitTest/kotlin/ly/david/musicsearch/shared/feature/history/HistoryListItemTest.kt
@@ -1,34 +1,9 @@
 package ly.david.musicsearch.shared.feature.history
 
-import coil.Coil
-import com.google.testing.junit.testparameterinjector.TestParameterInjector
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.test.UnconfinedTestDispatcher
-import kotlinx.coroutines.test.resetMain
-import kotlinx.coroutines.test.setMain
-import ly.david.musicsearch.ui.test.image.FakeImageLoader
 import ly.david.musicsearch.ui.test.screenshot.ScreenshotTest
-import org.junit.After
-import org.junit.Before
 import org.junit.Test
-import org.junit.runner.RunWith
 
-@OptIn(ExperimentalCoroutinesApi::class)
-@RunWith(TestParameterInjector::class)
 class HistoryListItemTest : ScreenshotTest() {
-
-    @Before
-    fun before() {
-        val fakeImageLoader = FakeImageLoader()
-        Coil.setImageLoader(fakeImageLoader)
-        Dispatchers.setMain(UnconfinedTestDispatcher())
-    }
-
-    @After
-    fun teardown() {
-        Dispatchers.resetMain()
-    }
 
     @Test
     fun releaseGroup() {
