+++ shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_artistWithCoverArt[NIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:0187ccaf1a0ad6880298c6a3c6021e0a5e70f5adadbdd64e9ac56f2b1a7fd215
-size 16478
+oid sha256:721dd7e70b7311f4eed9ce868684413579e2aa04ce6c8dba8b5caed94d576dc4
+size 16605
