+++ shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_artist[NIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:9ce53d15aa475a45108fcba7617ab8170a4f64c7c8f8e9a9623ab720790aa02d
-size 14524
+oid sha256:e35ff71327eef4ebec5aa8a5ccf33c50e5492245a441cfc8b8844b04ecf6ff27
+size 14644
