+++ shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_releaseGroupWithCoverArt[NOTNIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:0a0b08a71eb94cc676f2165d5ed890b3932a8fe466fe57d74410b6a552fa9b1d
-size 25920
+oid sha256:2efe91d213adc28d810293b42e1ae1f5e1b62da3bf1d7a876ced87ff8a5a411b
+size 26023
