+++ shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_releaseWithCoverArt[NOTNIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:536f6863daa0608d017995cb01487f8a58bfff9e9e05a04ac61eae96f3a93e40
-size 22096
+oid sha256:5efdb23c4cab8c7d34c46dacfc50b6545da0eee18c7b85bd58a83691e92763f6
+size 22196
