+++ shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_release[NIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:7b0da35b828d48fb08af7114828a5eab05672da5f262cb2e67ac62f94f98112c
-size 25645
+oid sha256:ccbfa387990255dcd92d6ae42a6ccd4d1363cbdd60bdaff87744ed3e4f8679b2
+size 25730
