+++ shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_release[NOTNIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:941c7e7568c36cb36612f224475cbb378b33937dba7fe038333472391c88339a
-size 25750
+oid sha256:7090b46f1f72f0a9aaf9fde24d26ad5c34b0842b184b5cf432006f2ca27a1117
+size 25860
