+++ ui/common/build.gradle.kts
@@ -56,11 +56,7 @@ kotlin {
         }
         val androidUnitTest by getting {
             dependencies {
-                implementation(projects.ui.test.image)
                 implementation(projects.ui.test.screenshot)
-                implementation(libs.bundles.kotlinx.coroutines)
-                implementation(libs.coil.compose)
-                implementation(libs.coil.test)
                 implementation(libs.test.parameter.injector)
                 implementation(libs.koin.test)
             }
