+++ ui/common/src/androidUnitTest/kotlin/ly/david/musicsearch/ui/common/release/ReleaseListItemTest.kt
@@ -1,32 +1,10 @@
 package ly.david.musicsearch.ui.common.release
 
-import coil.Coil
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.test.UnconfinedTestDispatcher
-import kotlinx.coroutines.test.resetMain
-import kotlinx.coroutines.test.setMain
 import ly.david.musicsearch.core.models.listitem.ReleaseListItemModel
-import ly.david.musicsearch.ui.test.image.FakeImageLoader
 import ly.david.musicsearch.ui.test.screenshot.ScreenshotTest
-import org.junit.After
-import org.junit.Before
 import org.junit.Test
 
-@OptIn(ExperimentalCoroutinesApi::class)
 class ReleaseListItemTest : ScreenshotTest() {
-
-    @Before
-    fun before() {
-        val fakeImageLoader = FakeImageLoader()
-        Coil.setImageLoader(fakeImageLoader)
-        Dispatchers.setMain(UnconfinedTestDispatcher())
-    }
-
-    @After
-    fun teardown() {
-        Dispatchers.resetMain()
-    }
 
     @Test
     fun releaseListItem() {
