+++ ui/common/src/test/snapshots/images/ly.david.musicsearch.ui.common.area_AreaListItemTest_releaseEventNoDate[NIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:1310fcc41688e9b4468896051e95362ccd9aa1938165cbd1d89604a51c4ac18b
-size 11148
+oid sha256:06057f80ca69ff8009dcdac33a90160c43f9aba9e64cecf8ea2a4574ad8fe25a
+size 11147
