+++ ui/common/src/test/snapshots/images/ly.david.musicsearch.ui.common.area_AreaListItemTest_releaseEventNoDate[NOTNIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:641330e7e457d1953e2f10f587dba4f1515c6f3105c974b3401d861e5d224272
-size 11337
+oid sha256:1f541e93047bfec1bf1deb6472b695da038fbb7aec8f0bf6865d4fb5e99d7b1f
+size 11360
