+++ ui/common/src/test/snapshots/images/ly.david.musicsearch.ui.common.artist_ArtistListItemTest_allInfo[NIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:fd84fe7ecb3ddf7b97974e9c90956806bdabac7b8c352f5631ae516a5ff577a3
-size 82465
+oid sha256:796763f9c62b24e99b55ada80c5f97609275a271285744fb77c282aa79048801
+size 82454
